services/cortex-agents/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-agents"
version = "0.3.0"
requires-python = ">=3.12"
dependencies = [
    "httpx==0.27.0",
    "pydantic==2.7.1",
    "nats-py==2.7.0",
    "ldap3==2.9.1",
    "dnspython==2.6.1",
    "cryptography==42.0.5",
    "structlog==24.1.0",
]
